Feature/tunning mode - Implementação do stream data

+++ modules/robot/services/CarStatusService/src/CarStatusService.cpp
@@ -23,27 +23,29 @@ CarStatusService::CarStatusService(std::string name, uint32_t stackDepth, UBaseT
 
     mappingService = MappingService::getInstance();
 
-    latMarks->marks->loadData();
-
-    if (latMarks->marks->getSize() <= 0)
+    if(!status->TunningMode->getData())
     {
-        status->encreading->setData(false);
-        status->robotIsMapping->setData(true);
-    }
-    else
-    {
-        status->robotIsMapping->setData(false);
-        status->encreading->setData(true);
-        numMarks = latMarks->marks->getSize();
-        mediaEncFinal = latMarks->marks->getData(numMarks - 1).MapEncMedia;
-        speed->setToCurve();
-    }
+        latMarks->marks->loadData();
 
+        if (latMarks->marks->getSize() <= 0)
+        {
+            status->encreading->setData(false);
+            status->robotIsMapping->setData(true);
+        }
+        else
+        {
+            status->robotIsMapping->setData(false);
+            status->encreading->setData(true);
+            numMarks = latMarks->marks->getSize();
+            mediaEncFinal = latMarks->marks->getData(numMarks - 1).MapEncMedia;
+        }
+    }
     status->robotState->setData(CAR_STOPPED);
 
     stateChanged = true;
     lastMappingState = false;
     lastState = status->robotState->getData();
+    lastTrack = (TrackState) status->TrackStatus->getData();
 
     firstmark = false;
 
@@ -87,7 +89,7 @@ void CarStatusService::Run()
     LEDsService::getInstance()->queueCommand(command);
     vTaskDelay(1500 / portTICK_PERIOD_MS);
     // Deletar o mapeamento caso o botão de boot seja mantido pressionado e exista mapeamento na flash
-    if(!gpio_get_level(GPIO_NUM_0) && latMarks->marks->getSize() > 0)
+    if(!gpio_get_level(GPIO_NUM_0) && latMarks->marks->getSize() > 0 && !status->TunningMode->getData() && status->HardDeleteMap->getData())
     {
         DataStorage::getInstance()->delete_data("sLatMarks.marks");
         status->encreading->setData(false);
@@ -99,27 +101,69 @@ void CarStatusService::Run()
     ESP_LOGD(GetName().c_str(), "Iniciando delay de 1500ms");
     vTaskDelay(1500 / portTICK_PERIOD_MS);
 
-    if (status->robotIsMapping->getData())
+    if (status->robotIsMapping->getData() && !status->TunningMode->getData())
     {
         ESP_LOGD(GetName().c_str(), "Mapeamento inexistente, iniciando robô em modo mapemaneto.");
         command.color = LED_COLOR_YELLOW;
         LEDsService::getInstance()->queueCommand(command);
         vTaskDelay(1000 / portTICK_PERIOD_MS);
         // Começa mapeamento
+        status->RealTrackStatus->setData(UNDEFINED);
+        status->TrackStatus->setData(UNDEFINED);
         mappingService->startNewMapping();
     }
 
-    status->robotState->setData(CAR_IN_CURVE);
+    if(!status->TunningMode->getData())
+    {
+        status->robotState->setData(CAR_IN_LINE);
+        if(!status->robotIsMapping->getData())
+        {
+            status->TrackStatus->setData(SHORT_LINE);
+            status->RealTrackStatus->setData(SHORT_LINE);
+        }
+        started_in_Tuning = false;
+    }
+    else
+    {
+        started_in_Tuning = true;
+        status->robotState->setData(CAR_TUNING);
+        status->TrackStatus->setData(TUNING);
+        status->RealTrackStatus->setData(TUNING);
+        status->encreading->setData(false);
+        status->robotIsMapping->setData(false);
+        latMarks->marks->clearAllData();
+        numMarks = 0;
+        mediaEncFinal = 0;
+        command.led[0] = LED_POSITION_FRONT;
+        command.led[1] = LED_POSITION_NONE;
+        command.color = LED_COLOR_WHITE;
+        command.effect = LED_EFFECT_SET;
+        command.brightness = 0.5;
+        LEDsService::getInstance()->queueCommand(command);        
+    }
     status->FirstMark->setData(false);
-
     // Loop
     for (;;)
     {
         vTaskDelayUntil(&xLastWakeTime, 100 / portTICK_PERIOD_MS);
-
+        
         status->stateMutex.lock();
+        TrackLen = (TrackState)status->TrackStatus->getData();
         pulsesBeforeCurve = latMarks->PulsesBeforeCurve->getData();
         pulsesAfterCurve = latMarks->PulsesAfterCurve->getData();
+        actualCarState = (CarState) status->robotState->getData();
+        if(started_in_Tuning && status->TunningMode->getData() && status->robotState->getData() != CAR_TUNING &&  !status->encreading->getData() && !status->robotIsMapping->getData()) 
+        {
+            status->robotState->setData(CAR_TUNING);
+            status->TrackStatus->setData(TUNING);
+            status->RealTrackStatus->setData(TUNING);
+            command.led[0] = LED_POSITION_FRONT;
+            command.led[1] = LED_POSITION_NONE;
+            command.color = LED_COLOR_WHITE;
+            command.effect = LED_EFFECT_SET;
+            command.brightness = 0.5;
+            LEDsService::getInstance()->queueCommand(command);  
+        }
         if(latMarks->rightMarks->getData() >= 1 && !firstmark)
         {
             firstmark = true;
@@ -134,34 +178,83 @@ void CarStatusService::Run()
             ESP_LOGD(GetName().c_str(), "Alterando velocidades para modo mapeamento.");
             command.color = LED_COLOR_YELLOW;
             LEDsService::getInstance()->queueCommand(command);
-            speed->setToMapping();
         }
 
-        else if (lastState != status->robotState->getData() && !lastMappingState && status->robotState->getData() != CAR_STOPPED)
+        else if ((lastState != status->robotState->getData() || lastTrack != (TrackState)status->TrackStatus->getData() || lastTransition != status->Transition->getData()) && !lastMappingState && status->robotState->getData() != CAR_STOPPED && status->robotState->getData() != CAR_TUNING)
         {
             lastState = status->robotState->getData();
-
-            if (lastState == CAR_IN_LINE)
+            lastTrack =  (TrackState)status->TrackStatus->getData();
+            lastTransition = status->Transition->getData();
+            if (lastState == CAR_IN_LINE && !lastTransition)
             {
-                ESP_LOGD(GetName().c_str(), "Alterando velocidades para modo inLine.");
-                speed->setToLine();
+                ESP_LOGD(GetName().c_str(), "Alterando os leds para modo inLine.");
                 command.led[0] = LED_POSITION_FRONT;
                 command.led[1] = LED_POSITION_NONE;
                 command.color = LED_COLOR_GREEN;
                 command.effect = LED_EFFECT_SET;
-                command.brightness = 0.5;
+                switch (TrackLen)
+                {
+                    case SHORT_LINE:
+                        command.brightness = 0.05;
+                        break;
+                    case MEDIUM_LINE:
+                        command.brightness = 0.3;
+                        break;
+                    case LONG_LINE:
+                        command.brightness = 1;
+                        break;
+                    case SPECIAL_TRACK:
+                        command.color = LED_COLOR_PURPLE;
+                        command.brightness = 0.05;
+                        break;
+                    default:
+                        command.color = LED_COLOR_WHITE;
+                        command.brightness = 1;
+                        break;
+                }
                 LEDsService::getInstance()->queueCommand(command);
             }
-            else
+            else if(lastState == CAR_IN_CURVE && !lastTransition)
             {
-                ESP_LOGD(GetName().c_str(), "Alterando velocidades para modo inCurve.");
+                ESP_LOGD(GetName().c_str(), "Alterando os leds para modo inCurve.");
                 command.led[0] = LED_POSITION_FRONT;
                 command.led[1] = LED_POSITION_NONE;
                 command.color = LED_COLOR_RED;
                 command.effect = LED_EFFECT_SET;
+                switch (TrackLen)
+                {
+                    case SHORT_CURVE:
+                        command.brightness = 0.05;
+                        break;
+                    case MEDIUM_CURVE:
+                        command.brightness = 0.3;
+                        break;
+                    case LONG_CURVE:
+                        command.brightness = 1;
+                        break;
+                    case ZIGZAG:
+                        command.color = LED_COLOR_PURPLE;
+                        command.brightness = 1;
+                        break;
+                    case SPECIAL_TRACK:
+                        command.color = LED_COLOR_PURPLE;
+                        command.brightness = 0.05;
+                        break;
+                    default:
+                        command.color = LED_COLOR_WHITE;
+                        command.brightness = 1;
+                        break;
+                }
+                LEDsService::getInstance()->queueCommand(command);
+            }
+            else if(lastTransition)
+            {
+                command.led[0] = LED_POSITION_FRONT;
+                command.led[1] = LED_POSITION_NONE;
+                command.color = LED_COLOR_ORANGE;
+                command.effect = LED_EFFECT_SET;
                 command.brightness = 0.5;
                 LEDsService::getInstance()->queueCommand(command);
-                speed->setToCurve();
             }
         }
 
@@ -179,31 +272,51 @@ void CarStatusService::Run()
 //         }
 //         iloop++;
 
-        if(!status->robotIsMapping->getData() && !status->encreading->getData()){
+        if(!status->robotIsMapping->getData() && !status->encreading->getData() && !status->TunningMode->getData() && actualCarState != CAR_STOPPED){
             robot->getStatus()->robotState->setData(CAR_STOPPED);
+            vTaskDelay(0);
+            DataManager::getInstance()->saveAllParamDataChanged();
+            command.led[0] = LED_POSITION_FRONT;
+            command.led[1] = LED_POSITION_NONE;
+            command.color = LED_COLOR_BLACK;
+            command.effect = LED_EFFECT_SET;
+            command.brightness = 1;
+            LEDsService::getInstance()->queueCommand(command);
         }
 
-        if (!status->robotIsMapping->getData() && actualCarState != CAR_STOPPED && status->encreading->getData() && firstmark)
+        if (!status->robotIsMapping->getData() && actualCarState != CAR_STOPPED && status->encreading->getData() && firstmark && (!status->TunningMode->getData() || !started_in_Tuning))
         {
             if ((mediaEncActual - initialmediaEnc) >= mediaEncFinal)
             {
-                ESP_LOGD(GetName().c_str(), "Parando o robô");
-                status->encreading->setData(false);
-                status->robotState->setData(CAR_IN_CURVE);
-                //vTaskDelay(100 / portTICK_PERIOD_MS);
+                if(status->TuningMapped->getData())
+                {
+                    status->FirstMark->setData(false);
+                    firstmark = false;
+                    initialmediaEnc = 0;
+                    status->robotState->setData(CAR_IN_LINE);
+                    status->TrackStatus->setData(SHORT_LINE);
+                    status->RealTrackStatus->setData(SHORT_LINE);
+                    latMarks->rightMarks->setData(0);
+                }
+                else
+                {
+                    ESP_LOGD(GetName().c_str(), "Parando o robô");
+                    status->encreading->setData(false);
+                    //vTaskDelay(100 / portTICK_PERIOD_MS);
 
-                // TODO: Encontrar forma bonita de suspender os outros serviços.
-                // vTaskSuspend(xTaskPID);
-                // vTaskSuspend(xTaskSensors);
+                    // TODO: Encontrar forma bonita de suspender os outros serviços.
+                    // vTaskSuspend(xTaskPID);
+                    // vTaskSuspend(xTaskSensors);
 
-                robot->getStatus()->robotState->setData(CAR_STOPPED);
-                DataManager::getInstance()->saveAllParamDataChanged();
-                command.led[0] = LED_POSITION_FRONT;
-                command.led[1] = LED_POSITION_NONE;
-                command.color = LED_COLOR_BLACK;
-                command.effect = LED_EFFECT_SET;
-                command.brightness = 1;
-                LEDsService::getInstance()->queueCommand(command);
+                    robot->getStatus()->robotState->setData(CAR_STOPPED);
+                    DataManager::getInstance()->saveAllParamDataChanged();
+                    command.led[0] = LED_POSITION_FRONT;
+                    command.led[1] = LED_POSITION_NONE;
+                    command.color = LED_COLOR_BLACK;
+                    command.effect = LED_EFFECT_SET;
+                    command.brightness = 1;
+                    LEDsService::getInstance()->queueCommand(command);
+                }
             }
             if ((mediaEncActual - initialmediaEnc) < mediaEncFinal)
             {
@@ -219,16 +332,20 @@ void CarStatusService::Run()
                     {
                         CarState trackType = (CarState)latMarks->marks->getData(mark+1).MapStatus;
                         TrackState trackLen = (TrackState)latMarks->marks->getData(mark+1).MapTrackStatus;
+                        status->RealTrackStatus->setData(trackLen);
+                        bool transition = false;
                         // Verifica se o robô precisa reduzir a velocidade, entrando no modo curva
                         if((CarState)latMarks->marks->getData(mark).MapStatus == CAR_IN_CURVE && (CarState)latMarks->marks->getData(mark + 1).MapStatus == CAR_IN_LINE)
                         {
                             if((Manualmedia + pulsesAfterCurve) < ManualmediaNxt && (mediaEncActual - initialmediaEnc) < (Manualmedia + pulsesAfterCurve)) 
                             {
+                                transition = true;
                                 trackType = CAR_IN_CURVE;
                                 trackLen = (TrackState)latMarks->marks->getData(mark).MapTrackStatus;
                             }
                             else if((Manualmedia + pulsesAfterCurve) >= ManualmediaNxt) 
                             {
+                                transition = true;
                                 trackType = CAR_IN_CURVE;
                                 trackLen = (TrackState)latMarks->marks->getData(mark).MapTrackStatus;
                             }
@@ -239,17 +356,20 @@ void CarStatusService::Run()
                             {
                                 if((ManualmediaNxt - pulsesBeforeCurve) > Manualmedia && (mediaEncActual - initialmediaEnc) > (ManualmediaNxt - pulsesBeforeCurve)) 
                                 {
+                                    transition = true;
                                     trackType = CAR_IN_CURVE;
                                     trackLen = (TrackState)latMarks->marks->getData(mark+2).MapTrackStatus;
                                 }
                                 else if((ManualmediaNxt - pulsesBeforeCurve) <= Manualmedia) 
                                 {
+                                    transition = true;
                                     trackType = CAR_IN_CURVE;
                                     trackLen = (TrackState)latMarks->marks->getData(mark+2).MapTrackStatus;
                                 }
                             }
                         }
                         // Atualiza estado do robô
+                        status->Transition->setData(transition);
                         status->robotState->setData(trackType);
                         status->TrackStatus->setData(trackLen);
                         break;
